changing GUI file name
